:art: Reduce duplication for 2/3 components
